+++ extcss3/utils.h
@@ -62,12 +62,12 @@
 
 /* ==================================================================================================== */
 
-unsigned short int extcss3_char_len(char c);
+unsigned int extcss3_char_len(char c);
 
 bool extcss3_ascii_strncasecmp(const char *str1, const char *str2, unsigned char n);
 
-char *extcss3_get_type_str(unsigned short int type);
-char *extcss3_get_flag_str(unsigned short int flag);
+char *extcss3_get_type_str(unsigned int type);
+char *extcss3_get_flag_str(unsigned int flag);
 
 /* ==================================================================================================== */
